Add logging support to MEX interface
Adds logging capabilities to the MATLAB MEX interface, i.e., the log
messages are forwarded to MATLAB via the mexPrintf() function. This
requires linking libcadet-mex against the necessary MATLAB libmex and
libmx libraries. A new logging target / destination, called MexOutput,
has been added and the corresponding setup of the logging framework is
done if the MATLAB_MEX_FILE preprocessor define is set. The CMake
scripts are adapted accordingly to set this define in MEX relevant
builds. The default error level in the MATLAB interface is raised to
ERROR in order to avoid a flood of irrelevant messages.

Closes #2.