Add second bound state support
Adds the support for a second bound state by introducing pseudo
components. Only the bound phase of the pseudo components is actually
used. Transport equations are solved on the other phases.
There are two variants:
A) Every component is doubled / gets a pseudo component
B) Every component but the first (salt) is doubled
Variant B) is useful for SMA-like binding models.