Add weights to fitColumn script
The fitColumn script has now the ability to apply weights to components
and experiments. If multiple components have been observed, then each
component can be weighted to, for example, improve the fit of small
peaks. When fitting multiple experiments simultaneously it is now
possible to apply a weight to each experiment.

Both weights can be used simultaneously and individually.