Fix bug in file-based backend of MATLAB interface
When creating a Simulator object in the MATLAB interface the version of
the underlying CADET simulator is retrieved. In case of the file-based
backend this fails with an error. In the file-based backend the version
is obtained by parsing the command line output of the cadet-cs binary.
The parsing code contained a wrong index which has been fixed.

Fixes #9.