Refactor parameter handling and improve performance
The parameters are moved from a std::map to dedicated storage in the
classes (GeneralRateModel, AdsorptionModels). This means that the
parameters are not obtained by a map lookup in each evaluation of the
residual. Apparently, this lookup was a major bottleneck which has been
resolved now. Consequently, performance is increased significantly.

The ParameterContainer class now stores pointers to the Parameter
objects instead of the objects themselves. Since the parameter values
are not longer obtained by querying the ParameterContainer, the class
is cleaned by removing such methods.