Add option to publish the git branch in binaries
Publishes the name of the current git branch in the binaries if the
corresponding option PUBLISH_BRANCH_INFO is set. This can be useful to
quickly distinguish between modified versions.

Publishing the branch name is disabled by default when building the
branch. It is enabled when building any other branch. The setting can
be overwritten by manually specifying it when calling cmake.