Remove write back of version to file in CMakeLists
After reading the version from the latest git tag in the history of the
current commit, the version was written to version.txt file. This can
cause trouble on Windows machines because of spaces in the path to the
version.txt file. The mechanism is now completely removed and people
need to update the version.txt file manually upon release of a new
version.