Fix bug in join error checking of fitColumn script
The error checking of joined parameters in the fitColumn script
calculated the total number of joined parameters sequentially in a loop
while checking each join for index errors. This is wrong since the
total number is needed in each loop.
The bug is fixed by calculating the total number of joined parameters
in one go before the loop starts.