Add meta group to file format
The meta group is added to the file format. It contains a field for the
file format version, which can be read in. However, this feature is
currently not used. The main purpose of the meta group is to hold
version info about CADET (version number, git SHA1, git branch).