Add resuming capabilities to simulator and interface
The full last state of solution and sensitivity systems can now be
returned. By reading the state back in, it is possible to resume a
simulation from the last state.

Flags for writing the last state vectors are added and the
documentation is updated accordingly. A new forward example
demonstrates the new resuming capabilities.