Add changes to enable batch resin uptake models
By allowing an interstitial velocity of zero, it is possible to
simulate batch resin uptake models. The column is degraded to a tank by
setting axial dispersion and interstitial velocity to zero. Simulations
can then be conducted with a minimum number of column cells (which is 3
at the moment due to the implementation of the WENO scheme). Since no
inlet condition is required any more, only one inlet section spanning
the whole computational domain is required. Note that the whole
solution must be written in order to get the solution inside the
particles. An example in example/forward/batchResinUptake.m is included.