Fix ignored number of threads
Due to the changes in commit 21a0c84821cc2f6c817ea5fbe7e9eca98a37fc9a
the number of threads was not read correctly. The reader is not in the
correct group which led the driver to assume that no number of threads
were set. At this point, the OpenMP default of using all available
threads (number of CPU cores) kicks in (except when overriden by the
environment variable OMP_NUM_THREADS). By directing the reader to the
right group just before reading the number of threads setting, this
issue is fixed.