Fix ignored WRITE_SENS_ALL setting
In commit 21a0c84821cc2f6c817ea5fbe7e9eca98a37fc9a a copy and paste
error has been introduced in the TimeIntegrator::writeSolution()
function. Instead of using the variables determining which parts of the
sensitivities should be stored, the ones for the original solution were
used. This is now fixed by changing them to the correct variables.