Fix missing root node in XML format
The XML format was invalid because of a missing root node. The input
and output groups would be placed side by side in the document, but the
standard does only allow one root node. This is fixed by adding an
appropriate root node where the input and output groups are placed in.
The new root node is named "cadet".