Replace ADOL-C with homegrown SFAD library
ADOL-C is replaced with a homegrown forward AD library. SFAD, which
stands for Simple Forward Automatic Differentiation, offers a simple
operator overloaded data type and an Expression Template approach.
Both variants have the possibility to store the derivative arrays in
dynamically allocated heap memory or on the stack (which is found to
be faster). However, using stack memory limits the number of usable
derivatives to some compile-time value.

ADOL-C is not removed from the repository, but the user can choose from
the available libraries at compile time via a CMake flag (-DAD_LIB).
This effectively removes the ADOL-C dependency from the project.

The code is further improved by reducing AD related loop iterations
as much as possible.