Release of version 3.0.0
CADET has been (almost fully) rewritten with a complete new OOP / code
architecture. The rewrite enables and eases the implementation of new
features in the near future and has improved stability as well as
performance of the simulation.

Major highlights:
- Completely new code architecture
- Increased single- and multi-core performance
- Improved documentation (quality and quantity)
- Improved MATLAB examples
- Preparations for upcoming features (networks of unit operations and
  support of reaction terms)
- Consistent initialization of the DAE solver
- Modernized CMake build system
- Native support of joined parameters via AD

Unfortunately, the changes are breaking, that is, users have to migrate
their code to the new version.