Fix failing lower WENO orders 1 and 2
The GeneralRateModel uses WENO for handling the convective term. There
are three WENO orders implemented: 1-1, 2-3, and 3-5. Only order 3-5
worked and a simulation using the other (lower) orders failed.

The problem has been traced back to the incorrect setup of the
CachingStencil in GeneralRateModel. The size of the stencil did not
adapt to the requirements of the WENO method but always assumed the
highest order.

This issue is now fixed.