Fix platform-dependent WENO order 1 bug (upwind)
While Linux and MacOS run just fine with the previous WENO fix, MSVC
binaries fail on Windows. The problem manifests by a seriously wrong
consistent initialization. It traces back to incorrect bandwidths of
the column bulk matrix blocks. The problem has been fixed by
guaranteeing that there is enough space in each row for handling the
dispersion terms (which was not the case for WENO 1-1 where the upper
bandwidth was accidentally set to 0).