Add NOMINMAX define to Windows builds
MS Visual Studio unfortunately uses headers that define min and max as
macros. Thus, using std::max() and std::min() fails. This issue is
fixed by defining NOMINMAX globally as a compiler flag. It had already
been in CMake as a target compile option, but has not come into effect.
Now a target compile definition is used to hand the definition to the
compiler.