Separate particle mobile phase and solid phase
Separates the notion of particle mobile phase and solid phase. The solid
phase is now recorded separately and instead of the whole particle,
which consists of mobile and solid phase, only its mobile phase is
saved. This allows to record the solid phase of CSTRs without hacks.

Adds WRITE_XYZ_SOLID in the /return/unit_XYZ group and XYZ_SOLID fields
in the /output subgroups.