Change bulk state ordering in GRM to cell-major
The ordering of the bulk DOF was component-major (i.e., comp0cell0, ...
comp0cellN, comp1cell0, ..., comp1cellN, ...). In order to prepare for
the introduction of reaction terms, this ordering is changed to cell-
major (i.e., cell0comp0, ..., cell0compN, cell1comp0, ..., cell1compN).
Instead of nComp band matrices for the bulk Jacobian that are solved in
parallel, there is now a single large band matrix for the column block.