Change FILE_FORMAT field in /meta group
Changes the data type of the FILE_FORMAT field in the /meta group of
the file format specification. The field used to be string, but is now
required to be an int. The version is divided into Major.Minor.Patch
and each part is assumed to have at most two digits. To form an int
from a version, the dots are ignored (e.g., 12.34.56 = 123456 or
3.0.0 = 30000). This obviates the need for parsing the version string.
Also pushes the FILE_FORMAT version from 3.0.0 to 3.1.0.