Register unit models via functions in ModelBuilder
Previously, unit operation models were registered by a template function
that required the type of the model. This required the inclusion of the
corresponding header file. In order to strengthen encapsulation, each
unit operation model now provides a function to register itself in the
factory map. This function is forward declared and later called in the
ModelBuilder class, which allows to get rid of the includes.