Fix problems in Matlab interface of column unitss
Fixes some problems with mandatory fields that should actually be
optional in the Matlab interfaces of the unit operations (GRM, LRM,
CSTR, LRMP). Also fixes a problem with interstitialVelocity being
optional now if crossSectionArea is set. In this case,
interstitialVelocity controls the flow direction via its sign. This
requires negative values to be allowed.

Also fixes a problem with PORE_ACCESSIBILITY in the GRM interface.