Expose nonlinear solver settings from IDAS
Exposes error control of forward sensitivity systems, number of Newton
iterations, number of convergence test failures, number of error test
failures, and number of forward sensitivity Newton iterations to the
user. The IDAS default values are used by default. The corresponding
settings in the file format are optional. Functionality is also exposed
to the Matlab interface.