Use CMake object library to avoid linking in tests
Avoids linking to libcadet in unit tests by using a CMake object
library. The object files belonging to libcadet are collected in an
object library that is reused for shared and static builds of libcadet
as well as in the unit test binary. This speeds up compilation as the
object files are only produced once.

This also fixes a problem on Windows where some unit tests required the
export of some symbols that were not exported by default. This issue
does not exist on Linux and Mac platforms as all symbols are exported
there.