Change comparison behavior of Approx in test
By introducing a custom type RelApprox, the behavior of comparisons with
Catch::Approx is changed. Instead of checking the absolute deviation
first, RelApprox checks the relative deviation. This was the standard
order in CATCH v1. Additionally, the updated relative comparison of
CATCH v2 is used. The scale factor is set 0.0 by default and the own
value is used as reference instead of the maximum absolute value of both
operands.