Fix configuration bug with too large NBOUND array
If the NBOUND array in the input file had more entries than components,
the _disc.nBound array would be written past its end. This means that
unallocated memory would be accessed which can cause harm in multiple
ways. This problem is fixed by checking the size and copying only what
is needed.