Pass memory buffer as parameter to binding models
Passes memory buffer to residual(), analyticJacobian(), and
timeDerivativeAlgebraicResidual() in binding models. This buffer can be
used for caching in externally dependent binding models.

Also changes name of consistentInitializationWorkspaceSize() to
workspaceSize(). This function now returns required workspace size in
bytes instead of number of doubles. This also includes the memory buffer
passed to the functions mentioned above.