Change ORDER_PP to custom template processor
Replaces the ORDER_PP C/C++ preprocessor interpreter with a custom
solution based on JSON and INJA. This was necessary as MSVC could not
handle ORDER_PP.

The new solution first builds a simple template processor called
templateCodeGen. This target has no dependencies except for the
header-only libraries INJA and JSON. INJA takes templates that are
similar to Python's JINJA template engine and populates them with data
from a JSON object. The ORDER_PP code is replaced with a template file
that generates all required code. Instead of invoking the macro in the
binding model, a special C++ comment with a JSON object is placed at
the top which serves as input for the template. templateCodeGen scans
the data file for the special comment blocks and replaces them with the
instantiated template.

Instead of compiling the binding model files directly, CMake is
instructed to run templateCodeGen on each file and compile the result
instead. The processed files are placed in the build tree.

JSON is updated to version 3.1.2 in order to comply with INJA's
requirements. ORDER_PP is removed from CADET's dependencies. Note that
INJA is only used to build templateCodeGen. INJA was chosen because it
is small, header-only, and obviates the need for another tool that has
to run under all platforms (e.g., Python).