Don't skip consistent init of sensitivities
Do not skip consistent initialization of sensitivity systems if their
full state vectors are given as initial values. The given values are not
trusted by default anymore. If consistent initialization is to be
skipped, because consistency of the given values is known, the simulator
needs to be ordered to do so explicitly by setting the consistent init
mode appropriately (preferred method) or by calling
skipConsistentInitialization().