Fix incorrect section index in external functions
Both external functions need to find an index in some array (usually
time points) determined by the current integration time and axial
position. The index calculation is incorrect as the std::lower_bound()
returns the index of the first element greater or equal the given time
instead of the last element smaller or equal the given point.

The bug is fixed by checking whether the element returned by
std::lower_bound() is greater than the given time point and decrementing
the index if it is.