Fix incorrect number of required AD dirs in CSTR
The number of required AD directions is calculated in requiredADdirs().
The CSTR model always returned the required AD directions for
automatically calculating the Jacobian even if analytic Jacobian was
enabled.

The number of required AD directions is now calculated based on whether
analytic Jacobians are enabled or not. This saves a few AD directions
whenever a CSTR is involved in the model.