Refactor configure() and reconfigure() methods
Up to now, unit operation models had a configure() and a reconfigure()
method. The configure() function would set up discretization-related
parameters and model parameters, whereas the reconfigure() would only
read in model parameters. reconfigure() could be called multiple times
while configure() could only be called once.

This pair of functions is split up into configureModelDiscretization()
and configure(). Taking the binding model interface as example,
discretization-related parameters are set in
configureModelDiscretization() and model parameters are set in
configure().

This entangles responsibilities. In order for this to work, the binding
model needs to configure its nonlinear solver in its
configureModelDiscretization() function. This is required since the unit
operation has to allocate memory for consistent initialization. The
memory required is partially determined by the needs of the nonlinear
solver.