Split up setup of initial conditions
Up to now, initial conditions were read in and applied in the same one
function. This mechanism is split up into reading the initial conditions
from configuration in readInitialCondition() and into applying them to
the state vector in applyInitialCondition().

This distinction is also made in the Simulator interface. That is,
applyInitialCondition() applies the initial conditions currently set and
setInitialCondition() reads initial conditions from configuration
without applying them.