Fix compiler error when using SUNDIALS OpenMP
When configured with SUNDIALS_PREFER_OPENMP, the simulator uses the
SUNDIALS NVECTOR module with OpenMP support. The number of threads for
each NVECTOR can be set explicitly in order to limit the number of
threads. This code contains an invalid variable name that causes the
compilation to fail. Since this code is only compiled conditionally on
SUNDIALS_PREFER_OPENMP, this bug has not surfaced until now as OpenMP
support is recommended by SUNDIALS only for problem sizes exceeding
0.1M unknowns.

The problem is easily fixed by using the correct variable name.
Fixes #11.