Fix buggy dense submatrix extraction from dense AD
Fixes an offset error in the extraction of a dense submatrix from a
Jacobian generated by dense AD seed vectors. This bug only happens when
using DenseJacobianExtractor. It surfaces in the consistent
initialization of algebraic binding model equations when using a CSTR.

The direction offset in the extraction did not account for the row
offset, which lead to the extracted matrix being all 0.