Fix consistent sensitivity init in CSTR model
The computation of consistent initial sensitivities in the
StirredTankModel overwrites the Jacobian _jac after each parameter.
Hence, the solution of the time derivative system is wrong due to _jac
not containing the correct matrix after the first iteration.

This is fixed by not overwriting the matrix _jac in the function
multiplyWithDerivativeJacobian(), which now directly performs the
matrix-vector multiplication (i.e., matrix free).