Fix lifetime bug with ActiveSimulationTime temps
When turning a SimulationTime into an ActiveSimulationTime with
toActive(), the temporarily created actives are destroyed when leaving
the function. This renders the references in the ActiveSimulationTime
invalid.

The only occurrence of this function (toActive()) is removed and local
variables are created that live as long as the required temporary
ActiveSimulationTime.