Fix bug when reading empty JSON array
When reading an empty JSON array that is expected to be of type bool or
int, the program crashes. This is due to the JsonParameterProvider
checking the type of the first entry of the array in order to convert
from bool to int or vice versa if necessary. However, this reads from
uninitialized memory or dereferences nullptr if the array is empty.

The bug is fixed by adding a check for empty arrays.