Multiplex flow rate parameters in ModelSystem
When a single connection is split up into multiple parts (e.g., for
reordering the components), only the AD directions and value of the
first part of the connection are influenced by setParameterValue() and
friends. Now, the values are multiplexed to all relevant flow rates.

Note that this is actually a non-issue as only the first part of the
connection is used for computations.