Add support for writing strided blocks in writers
Adds support for writing blocks of data that may be strided in memory.
That is, blocks of data do not have to be contiguous, but can be placed
at regular memory positions (with a constant distance to each other,
which is called stride).

Also fixes some minor bugs when writing certain data. These bugs have
not shown up as such data is never written (e.g., arrays of bools,
arrays of strings).