Add multi-port unit support in tests and fix bug
Adds support for unit operations with multiple ports in the tests.

Also fixes a bug that has lead to a lot less comparisons in
testFwdSensSolutionFD() and testFwdSensSolutionForwardBackward() than
actually necessary. Only a small part of the results were compared
instead of the full range.

Note that this bugfix will make tests break as the error tolerances
and FD step sizes need to be adjusted to account for the increased
number of comparisons.