Fix floating point bug with USER_SOLUTION_TIMES
Under certain circumstances, the transformed USER_SOLUTION_TIME is
slightly above the current SECTION_TIME boundary. This causes IDA to
fail due to an incorrect stopping time. This behavior is caused by a
floating point error in the linear transformation of the time points,
although the transformation is perfectly fine, mathematically.

The problem is fixed by explicitly checking for the time point being
exactly one of the section times and returning an exact value instead of
computing the linear transformation.