Fix non-existent inlet / outlet bug in SolRecorder
If storeOutlet is enabled, but the unit operation assigned to the
solution recorder does not have an outlet, the recorder still tries to
write the dataset. As there is no data, this operation fails.

This problem is fixed by checking whether the unit operation actually
has an outlet and modifying the storeOutlet flag accordingly.

The same goes for inlets and the storeInlet flag.