Use Matlab's UMFPACK in MEX interface
On some platforms, the UMFPACK libraries provided by the user clash with
Matlab's own UMFPACK libraries. Since Matlab's symbols are already
loaded when the MEX interface is loaded, Matlab's UMFPACK is used. As
Matlab uses a custom build of UMFPACK with another integer type for
indexing, CADET crashes (when using 2D GRM and UMFPACK linear solver).

On the other hand, this also means that the user does not need to have
UMFPACK installed, but that we can use Matlab's UMFPACK for the MEX
interface.

For the cadet-cli binary, the status-quo is left unchanged, that is,
SuperLU, UMFPACK and dense linear solvers can be used (if available).
This still requires the user to provide SuperLU and UMFPACK libraries.

For the MEX interface, Matlab's UMFPACK is used as sparse direct solver.
SuperLU cannot be used sincce Matlab's UMFPACK requires a different
integer type for indexing. Hence, UMFPACK and dense solvers are
available for the MEX interface and no additional dependency is needed.
This means, the user does not need to have UMFPACK installed in order to
use it in the MEX interface.