Multiplex particle type and radial parameters
Adds multiplexing to parameters related to particle types. This means
that these parameters now also take less values. The given values are
then multiplexed to the full particle-type-dependent size. For example,
PAR_RADIUS has to be specified for each particle type. Now, PAR_RADIUS
can be given for all particle types (just one value) or it is specified
per particle-type (as before).

The same goes for parameters that depend on particle type and component.

In the 2D GRM, multiplexing for radially-dependent parameters is added
as well.