Fix Matlab serialization of multi binding models
When saving a model or simulator in Matlab and loading it back from
file, the binding models were not loaded correctly and an error was
thrown. This was caused by failure to assign a binding model to an
empty array of type double. Creating an empty array of type BindingModel
fixes the problem.