Add thread local storage based on thread index
Adds a class that provides thread local storage of an array based on the
index (or id) of the current thread. The maximum number of threads is
used to allocate an array of fixed size for each thread. The arrays are
placed on separate cache lines to avoid false sharing. Using the index
of the current thread, the corresponding array can be retrieved.

This is an alternative to tbb::enumerable_thread_specific. Since threads
may be spawned and removed by TBB at will, enumerable_thread_specific
could possibly allocate more arrays than the maximum amount of threads.
Also, each birth / death of a thread requires allocating / deallocating
its corresponding array. This inconvenience is, hopefully, circumvented
by our ThreadLocalStorage class.

If parallelization is disabled, only a single array is maintained.