Add support for thread local storage to models
The ModelSystem provides a thread local storage for the whole system. It
queries its unit operations for the size of the storage and allocates
the maximum size requested. The storage is then passed to the units for
certain operations.

Each unit operation uses the thread local storage provided by the
ModelSystem. This saves memory since we don't need to allocate enough
memory such that threads do not get in each other's way. The savings may
be substantial for larger models.