Remove storage policy from S(ET)FAD
Removes the possibility to choose between heap and stack storage in the
AD libraries SFAD and SETFAD. Stack storage is now hardwired into the
code. Heap storage is very slow when not combined with a suitable
allocator. Since we do not provide a thread-safe allocator that is able
to handle AD types, the heap storage option is removed.

Additionally, this allows to use default copy and assignment operators
and constructors, which make the AD type TriviallyCopyable. This, in
turn, may enable some compiler optimizations (memcpy instead of loop).