Rename SAI_KA to SAI_KA1 in accordance with docs
The documentation and Matlab interface does not match the code with
respect to the parameter SAI_KA. The documentation and Matlab interface
declare the name SAI_KA1 instead of the name SAI_KA1, which is
actually read in the code.

The code is changed to use SAI_KA1 instead of SAI_KA such that the
documentation is correct.