Improve types involving AD in binding models
In the binding models, variables of parameter and state type often need
to be combined in calculations. If one of them uses AD, the result also
is of active (AD) type. Up to now, the Residual type has been used,
which depends on parameter, liquid state, and solid state type. When
liquid state and parameter values are combined, a dedicated type is used
instead of the Residual type. This allows minor savings due to using
double instead of active if only the solid state type is active.