Fix backwards incompatible output naming scheme
With version 3.2.0, multi port unit operations have been introduced. In
addition, the output naming scheme has changed, that is, a _PORT_XZY_
part has been added to the name of the output fields. This was a
breaking change in the file format.

Setting /input/return/SPLIT_PORTS_DATA to 0 is a possible workaround,
but this results in 2D arrays (with one dimension of size 1) instead of
1D arrays, which may break code.

The old naming scheme is restored with the following changes:
For unit operations that only have a single inlet / outlet port, the
_PORT_XYZ_ part is removed from the field name. This produces the old
field names. The new behavior is requested by setting the newly
introduced field /input/return/SINGLE_AS_MULTI_PORT. If set, single port
unit operations are treated as multi port unit operations. The flag is
disabled by default.

In addition, output datasets of single port unit operations are
simplified (i.e., 2D arrays decay to 1D arrays if one dimension has size
1). This only happens if SINGLE_AS_MULTI_PORT is disabled, though.

Closes #42.