Accept empty bindingModel in CSTR of Matlab iface
The StirredTankModel in the Matlab interface does not accept an empty
bindingModel. This was possible up to (including) version 3.1.2.

Using a NoneBinding object for bindingModel does not work either since
it complains about missing bound states (even if nBoundStates = 0).
Adding a bound state crashes the simulation (since NoneBinding, which
maps to DummyBinding in C++, does not add equations for the bound
states).

The problem is fixed by adapting the buggy check in the
assembleConfig() function. Fixes #44.