Add sequential linear solver to ModelSystem
Adds a sequential method to solve the linear systems in ModelSystem. The
method can either be chosen by setting the flag LINEAR_SOLUTION_MODE, or
is detected automatically using a (very simple) heuristic (default).

The sequential solution method can only be used for graphs without
cycles. If a cycle is detected, the parallel method is used.