Fix reading beyond array in SlicedVector
When the last slice in SlicedVector is empty and the slice is accessed
by calling operator[] or at(), the _values array is accessed one item
beyond its size. The actual value is not used, though, since only the
address of the item is returned. The problem is fixed by using pointer
arithmetics instead of accessing the item to obtain its address.