Fix sensitivity problems with mass action reaction
Fixes a problem that leads to NaNs when computing the derivative of the
residual in the mass action law dynamic reaction model. The problem is
caused by state variables being slightly negative (within the tolerance
of the time integrator). The derivative of the power-law requires the
logarithm of the base, which results in NaN for negative base values.
The problem is fixed by checking the base value before calculating the
power.

In the context of the problem, the slightly negative values were caused
by consistent initialization. Ideally, these checks should not be
necessary in the reaction model. Instead, consistent initialization
should take care of providing nonnegative values.