Fix DPFR with reactions bug (LRM w/o bound states)
Fixes a bug that occurs when reactions are added to a DPFR model (or
LumpedRateModelWithoutPores and without bound states). When adding the
Jacobian of the reaction terms to the solid phase, we encounter an
error since there is no solid phase in a DPFR. This is prevented by an
if statement.