Add colloidal binding model
Adds a colloidal binding model developed by researchers of A. Lenhoff's
group. The binding model requires linearization at the origin, which is
controlled by a threshold. It uses a mobile phase modifier (comp. 0)
and, optionally, pH provided as pseudo component (comp. 1), which has
to be enabled by a flag.

The analytic Jacobian of this binding model is not yet fully verified.
If problems arise, use AD for computing the Jacobian (i.e., disable
USE_ANALYTIC_JACOBIAN).