Assign and fix column outlet based on flow direction
Depending on flow direction, the last (forward) or first (backward)
axial cell is selected as outlet of a column unit operation. In the 2D
model, the outlet is chosen per port based on its flow direction. This
fixes #56.

Also fixes a bug that caused incorrect behavior on flow reversal in the
middle of a simulation. Although the column units correctly report the
index of the outlet via localOutletComponentIndex(), this information
was queried at the wrong time in ModelSystem (i.e., before the
velocities of the models had been updated).