Avoid advancing uninitialized vector iterator
If the user does not provide output time points, the iterator for these
points is not initialized since the corresponding vector is empty.
Nevertheless, the iterator is advanced during time integration. This
triggers an assert in the MSVC standard library in debug mode.

We prevent this problem by placing the iterator advancement in an if
clause. Apparently, the standard libraries of GCC and clang do not
check for this scenario or do not consider it problematic.