Convert nu to piecewise cubic poly in GIEX model
The characteristic charge nu is extended to a piecewise cubic polynomial
in the GIEX adsorption model. The nu of each component can have
different polynomial coefficients and breaks, but their amount must be
the same over all components.

If the breaks are not provided, a single global polynomial is applied.
This maintains backwards compatibility.