Fix handling of flow direction for dynamic flow rates
If dynamic flow rates were enabled, the column velocity (which is used to determine the flow direction if the cross section area is also provided) was not taken into account.

`_curVelocity` was correct in the call to `residual()` after `notifyDiscontinuousSectionTransition()`. However, it was overwritten in subsequent calls to `residual()` which happened almost instantly (in terms of time) if dynamic flow rates are configured.