Increase linear solution mode heuristic barrier
When automatically selecting the linear solution mode of a model
system, the heuristic is solely based on the number of unit operations.
Of course, the parallel approach is chosen if any cycle is present in
the system.

If the system consists of at least 6 unit operations, parallel solution
mode is selected. Based on observations, this number is far too low in
practice. Hence, the barrier is bumped up to 25 unit operations.

Closes #103.