Fix missing CSTR bulk solution export (#150)
Fix missing CSTR bulk solution export

Because of the primary coordinate, the shape of the written bulk
solution will be (numTimesteps, 1, numComponents). This is a breaking
change compared to previous versions. In order to selectively remove the
singleton dimension, the function hasPrimaryExtent() is added to the
ISolutionExporter interface. This flag determines whether the unit
operation model actually extents in the primary coordinate direction,
that is, whether it is a true spatial element. If a unit operation model
does not have primary extent, the singleton dimension will be removed.