Automatically compute binding model Jacobians via AD if analytical Jacobian is not provided (#220)
* Automatically compute binding model Jacobians via AD

Defaults to automatically computing the Jacobian of a binding model if
it does not provide an analytical one. The implementsAnalyticJacobian()
is changed to a protected function that binding model implementations
need to implement. This determines whether AD will be used to compute
the Jacobian.
Note that the Jacobian will only be computed via AD for the binding, wherease the transport Jacobian is still computed analytically. This significantly reduces computational overhead compared to setting ``USE_ANALYTIC_JACOBIAN`` to true for the whole unit operation. 

Tests co-authored by Jan Breuer
---------

Co-authored-by: Jan Breuer <j.m.br@gmx.de>